+++ docs/changelog/changelog.md
@@ -9,7 +9,7 @@
 Version [`rva_17.101?`](https://github.com/Yethiel/re-volt-addon/releases/tag/rva_17.101?)
 
 * **Additions**
-  * **Texture Animation**: Added a transform animation feature (animates from frame A to frame B).
+  * **Texture Animation**: Transform animation feature (animates from frame A to frame B), Grid animation feature (for creating animations like the mars animation found in Museum 2)
   * **Re-Export Button**: Press the button next to the Export button to export to the same file again.
 
 * **Fixes**
@@ -17,6 +17,8 @@ Version [`rva_17.101?`](https://github.com/Yethiel/re-volt-addon/releases/tag/rv
 
 * **Misc.**
   * Added icons to panels
+  * Moved add-on settings to a separate panel
+  * Fixed the import/export panel to the top
   * Notifications now have an icon that matches the messages
   * Code refactoring